new Id returned

+++ core/src/main/java/de/itemis/recordstore/RecordRepository.java
@@ -1,5 +1,5 @@
 package de.itemis.recordstore;
 
 public interface RecordRepository {
-    Integer save(Record record);
+    Long save(Record record);
 }
